BUG-1696 : Consolidate test
